클라이언트는 HTML 파일의 __INITIAL_DATA__를 활용하여 렌더링하도록 변경
